complete logging work to check flow

+++ src/main/java/com/mentorship/hanakoleh/domain/order/exception/OrderNotOwnedByRestaurantException.java
@@ -1,6 +1,6 @@
 package com.mentorship.hanakoleh.domain.order.exception;
 
 public class OrderNotOwnedByRestaurantException extends RuntimeException {
-    public OrderNotOwnedByRestaurantException(Integer activeOrderId, Integer authenticatedRestaurantId) {
+    public OrderNotOwnedByRestaurantException(Long activeOrderId, Integer authenticatedRestaurantId) {
     }
 }
